Use standard assertEquals in TestUtil.compare
This way it prints the actual strings it compares - previously it
compared one pair of string, then printed a different pair on failure!

The default failure message is sometimes not as nice, but it highlights
the differences, and IDEs can show a better diff. If you need to see the
full strings, you can add prints locally.